requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
jinja2==3.1.6
